+++ src/helpers/analytics/trackGiveEvent.ts
@@ -1,0 +1,23 @@
+import { getGiveProjectName } from "../GiveProjectNameHelper";
+import { trackGAEvent } from "./trackGAEvent";
+
+export interface IUAData {
+  address: string;
+  value: string;
+  recipient: string;
+  approved: boolean;
+  txHash: string | null;
+  type: string;
+}
+
+export const trackGiveEvent = (uaData: IUAData, eventAction?: string) => {
+  trackGAEvent({
+    category: "Olympus Give",
+    action: eventAction ? eventAction : uaData.type ? uaData.type : "unknown",
+    label: getGiveProjectName(uaData.recipient, "unknown"),
+    value: Math.round(parseFloat(uaData.value)),
+    metric1: parseFloat(uaData.value),
+    dimension1: uaData.txHash ?? "unknown",
+    dimension2: uaData.address,
+  });
+};
